requirements.txt
Django==5.0.2
djangorestframework==3.14.0
djongo==1.3.6
pymongo==3.12.3
django-cors-headers==4.3.1
python-dotenv==1.0.1 